Made check for "addEventListener" method precise.
ActionScriptTypeChecker complains if a method "addEventListener" is
called with a callback that does not have a single parameter of type flash.events.Event
or starling.events.Event.
However, the check used to simply compare the method name, not whether it
actually refers to flash.events.IEventDispatcher#addEventListener() or
starling.events.EventDispatcher#addEventListener(). Thus, if a method
in another framework is named "addEventLister", but does not require the same
contract, you get false negatives.

This fix checks whether the object on which addEventListener() is invoked is of
a type that implements Flash's IEventDispatcher interface or extends Starling's
EventDispatcher class.

The test class StarlingEvent.as has been extended to mock Starling's EventDispatcher
class and to test the case that when another framework defines its own event
dispatcher (here: com.acme.MyEventDispatcher), no warning appears.