Support using constants for resource bundle names.
So far, IDEA only supports resource bundle references when given directly as string literals.
This changes additionally allows variables with a string literal initializer to be used.
This makes code more readable if fully-qualified resource bundle names are rather long.
Example:

  var foo:String = resourceManager.getString("my.very.long.package.MyBundle", "foo");
  var bar:String = resourceManager.getString("my.very.long.package.MyBundle", "bar");

is now also fully supported when introducing a constant:

  private static const MY_BUNDLE:String = "my.very.long.package.MyBundle";

  var foo:String = resourceManager.getString(MY_BUNDLE, "foo");
  var bar:String = resourceManager.getString(MY_BUNDLE, "bar");

Before, the second code would suggest / accept any key defined in any resource bundle,
not only the keys of MyBundle.

The adapted test (ResourceReferences2.mxml) proves that: The incorrect access to
key "prompt4" was not detected without the change to FlexPropertyReferenceProvider
(FlexHighlightingTest#testResourceReferences2() red), but after applying the changes
to FlexPropertyReferenceProvider, the test is green again.